requirements.txt
python-dotenv
flask
mysql-connector-python
openai